# tsc(perf): improve icon tsc perf