--- a/static/app/icons/iconAsana.tsx
+++ b/static/app/icons/iconAsana.tsx
@@ -2,10 +2,7 @@ import * as React from 'react';
 
 import SvgIcon, {SVGIconProps} from './svgIcon';
 
-const IconAsana = React.forwardRef<SVGSVGElement, SVGIconProps>(function IconAsana(
-  props: SVGIconProps,
-  ref: React.Ref<SVGSVGElement>
-) {
+const IconAsana = React.forwardRef<SVGSVGElement, SVGIconProps>((props, ref) => {
   return (
     <SvgIcon {...props} ref={ref}>
       <path d="M8,.61A3.48,3.48,0,1,1,4.52,4.09,3.48,3.48,0,0,1,8,.61Z" />