--- a/static/app/icons/iconBell.tsx
+++ b/static/app/icons/iconBell.tsx
@@ -1,11 +1,9 @@
 import * as React from 'react';
 
-import SvgIcon from './svgIcon';
+import SvgIcon, {SVGIconProps} from './svgIcon';
 
-type Props = React.ComponentProps<typeof SvgIcon>;
-
-const IconBell = React.forwardRef(function IconBell(
-  props: Props,
+const IconBell = React.forwardRef<SVGSVGElement, SVGIconProps>(function IconBell(
+  props: SVGIconProps,
   ref: React.Ref<SVGSVGElement>
 ) {
   return (