--- a/static/app/icons/iconCircle.tsx
+++ b/static/app/icons/iconCircle.tsx
@@ -2,10 +2,7 @@ import * as React from 'react';
 
 import SvgIcon, {SVGIconProps} from './svgIcon';
 
-const IconCircle = React.forwardRef<SVGSVGElement, SVGIconProps>(function IconCircle(
-  props: SVGIconProps,
-  ref: React.Ref<SVGSVGElement>
-) {
+const IconCircle = React.forwardRef<SVGSVGElement, SVGIconProps>((props, ref) => {
   return (
     <SvgIcon {...props} ref={ref}>
       <path d="M8,16a8,8,0,1,1,8-8A8,8,0,0,1,8,16ZM8,1.53A6.47,6.47,0,1,0,14.47,8,6.47,6.47,0,0,0,8,1.53Z" />