--- a/static/app/icons/iconClock.tsx
+++ b/static/app/icons/iconClock.tsx
@@ -1,11 +1,9 @@
 import * as React from 'react';
 
-import SvgIcon from './svgIcon';
+import SvgIcon, {SVGIconProps} from './svgIcon';
 
-type Props = React.ComponentProps<typeof SvgIcon>;
-
-const IconClock = React.forwardRef(function IconClock(
-  props: Props,
+const IconClock = React.forwardRef<SVGSVGElement, SVGIconProps>(function IconClock(
+  props: SVGIconProps,
   ref: React.Ref<SVGSVGElement>
 ) {
   return (