--- a/static/app/icons/iconCommit.tsx
+++ b/static/app/icons/iconCommit.tsx
@@ -1,11 +1,9 @@
 import * as React from 'react';
 
-import SvgIcon from './svgIcon';
+import SvgIcon, {SVGIconProps} from './svgIcon';
 
-type Props = React.ComponentProps<typeof SvgIcon>;
-
-const IconCommit = React.forwardRef(function IconCommit(
-  props: Props,
+const IconCommit = React.forwardRef<SVGSVGElement, SVGIconProps>(function IconCommit(
+  props: SVGIconProps,
   ref: React.Ref<SVGSVGElement>
 ) {
   return (