Icon changes taken from https://github.com/getsentry/sentry/pull/30847 ([perf review doc](https://www.notion.so/sentry/Typescript-performance-audit-ac375cf3d9374e66998a6eab2d5539e4))

Before:
```
Files:              4044
Lines:            652029
Nodes:           2506288
Identifiers:      816308
Symbols:         3258845
Types:           1011716
Instantiations:  3300586
Memory used:    2909886K
I/O read:          0.58s
I/O write:         0.00s
Parse time:        4.14s
Bind time:         1.30s
Check time:       33.94s
Emit time:         0.00s
Total time:       39.38s
```
<img width="991" alt="CleanShot 2022-01-11 at 12 39 11@2x" src="https://user-images.githubusercontent.com/9317857/148936238-ba9c62fd-f730-4017-a0d1-de30991d082f.png">


After:
```
Files:              4044
Lines:            651840
Nodes:           2506222
Identifiers:      816229
Symbols:         2641917 ↓ ~600k
Types:            875830
Instantiations:  3084392 ↓ ~300k
Memory used:    2439648K ↓ ~500MB 
I/O read:          0.30s
I/O write:         0.00s
Parse time:        2.92s ↓ ~1s
Bind time:         0.90s 
Check time:       27.88s ↓ ~6s
Emit time:         0.00s
Total time:       31.71s ↓ 7.67s decrease
```

<img width="1045" alt="CleanShot 2022-01-11 at 12 39 06@2x" src="https://user-images.githubusercontent.com/9317857/148936294-e32ace39-8d7f-405b-a4bd-76420efcf9a7.png">